Fix Android build issue.
